* IndexBuffer uses 16bit indices, (Wii does not support 32bit indices anyways). Note: Interface only changed on IndexBuffer::GetData(), if you are accessing IndexBuffers via operator[] there are no changes necessary * fixed renderer statistics bug when ResetDevice()/DestroyAll() was issued

+++ trunk/Wire/Engine/WireIndexBuffer.cpp
@@ -20,11 +20,11 @@ IndexBuffer::IndexBuffer(UInt quantity, UsageType usage)
 	Buffer(usage),
 	mQuantity(quantity)
 {
-	mpIndices = WIRE_NEW UInt[mQuantity];
+	mpIndices = WIRE_NEW UShort[mQuantity];
 }
 
 //----------------------------------------------------------------------------
-IndexBuffer::IndexBuffer(UInt* pIndices, UInt quantity, UsageType usage)
+IndexBuffer::IndexBuffer(UShort* pIndices, UInt quantity, UsageType usage)
 	:
 	Buffer(usage),
 	mQuantity(quantity),
@@ -38,8 +38,8 @@ IndexBuffer::IndexBuffer(const IndexBuffer* pIndexBuffer)
 	Buffer(pIndexBuffer->GetUsage()),
 	mQuantity(pIndexBuffer->mQuantity)
 {
-	mpIndices = WIRE_NEW UInt[mQuantity];
-	size_t size = mQuantity*sizeof(UInt);
+	mpIndices = WIRE_NEW UShort[mQuantity];
+	size_t size = mQuantity*sizeof(UShort);
 	System::Memcpy(mpIndices, size, pIndexBuffer->mpIndices, size);
 }
